Fix problems pairing "Security mode 3" devices
The PIN question would appear first on the device, and
the user would have to guess the PIN code asked by the
wizard to be able to pair, as we'd only show the PIN
after a PIN callback, which would come after the PIN
was asked on the device side.

Fixes pairing with Blackberries and Samsung phones.

(Closes: #578991)