Bug 583516 – When restarting bluetoothd, page doesn't go to adapter
Fix the setting of the default adapter in the adapter struct.
As we might be getting info about the adapter being the default
before the structure is created, check in the create_adapter
function instead of when allocating the structure, and use
adapter_changed to set it appropriately.

If we had decent dbus bindings, all this should be sync, and we
wouldn't have to create the adapter struct in an idle.