Fix crash when first killswitch is added after startup
If the applet starts with no killswitches installed, the following code
in main.c destroys the killswitch object:

if (bluetooth_killswitch_has_killswitches (killswitch) == FALSE)
{
	g_object_unref (killswitch);
	killswitch = NULL;
}

However, the destructor for the killswitch object
(bluetooth_killswitch_finalize) fails to cleanup the io_watch set in its
constructor. So when a killswitch is connected later, the event_cb for
the killswitch object runs, and attempts to use the now-destructed
killswitch object. The result is a crash.

Moreover, just because we don't have a killswitch at startup doesn't
mean we won't have one later.

This patch corrects both issues, by fixing the finalizer, and removing
the conditional destruction of the killswitch in main.c.

Signed-off-by: Bryan Donlan <bdonlan@gmail.com>

https://bugzilla.gnome.org/show_bug.cgi?id=598181