Set a parent on the remove confirmation dialog
With a NULL parent on Moblin the dialog ends up stacked below other windows
and can't accept any user input.
Modify the internal method show_confirm_dialog to take a BluetoothChooser
as an argument and set the toplevel window of the chooser as a parent for
the dialog.

This makes things work *much* nicer in Moblin and doesn't seem to have any
adverse affect in Gnome.

https://bugzilla.gnome.org/show_bug.cgi?id=598685