Fix crash when obex-client isn't running already
When obex-client isn't running yet, we can't get a name owner
for the service. But we'd still get a callback from the service
when trying to use it, and crash trying to compare the sender
with a NULL string.

Instead, set the owner's dbus name if we don't already have one
in obex_agent_request().

See: https://bugzilla.redhat.com/show_bug.cgi?id=544881