Use BluetoothClient for tracking the default-adapter
BluetoothClient already does all the work of tracking what adapter
is the default, and whether it is powered or not.
There is no need to duplicate that, and also we avoid searching
the default adapter more often than needed, with the side effect
of unregistering the agent (possibly in the middle of a transaction,
making it just fail).

https://bugzilla.gnome.org/show_bug.cgi?id=635023