wizard: make it possible to choose a random PIN after choosing a fixed one
Do not overwrite the user PIN with the random PIN and vice versa and generate
the random PIN only when it's needed; this also means that restarting the
pairing process doesn't silently overwrite user choices.
Also rewrite the generating function without side effects and fix the upper
limit of g_random_int_range() which is open.