wizard: fix again the choice of the automatic PIN
Give a value for every "pin" data to remove the ambiguity between NULL
PIN code and "pin" data not set, else random and custom PIN will be
swapped in the dialog. Now a NULL pointer means "use a random PIN" both
in the PIN database and in the GUI, while an empty string means "use
the contents of the text entry".