applet: Fix new pairings not appearing in menus
As documented now, the GtkTreeModelFilter that we use to
filter per-adapter devices lacks the ability to tell us
about row-changed signals (amongst others).

So just monitor those signals on the child-model of the filtered
model, and check whether those rows are in our view to trigger
the "devices-changed" signal.

https://bugzilla.gnome.org/show_bug.cgi?id=666127

Conflicts:

	applet/bluetooth-applet.c