wizard: fix pairing for devices with a pin in the DB
As far as I understand the intention of the code here is to set
automatic_pincode to FALSE if the DB detects it needs a special code
(e.g. keyboards or joysticks); otherwise if a pincode is specified in
the DB, that one should be used instead of being discarded.

https://bugzilla.gnome.org/show_bug.cgi?id=672077