sendto: keep the "Retry" button always visible and make it sensitive when needed
It is less surprising for the user if it is visible from the start instead of
appearing after an error and pushing the "Cancel" button to the left.
It also follows the advice in GNOME HIG:
"It is usually better to make a control insensitive than to hide it altogether.
This way, the user can learn about functionality they may be able to use later,
even if it is not available right now."
http://developer.gnome.org/hig-book/3.0/controls-sensitivity.html.en