wizard: Fix Simple Pairing never succeeding
The bluez RequestConfirmation call expect either an empty message (on
success) or an error (on failure) as return. The code tries to return
a string on success though, contrary to what the internal
introspection data says causing gdbus to just throw a warning
and bluez never getting a response, causing pairing to fail.

Fix RequestConfirmation return value

https://bugzilla.gnome.org/show_bug.cgi?id=679190