bluetooth-client: Port to BlueZ 5
Port the client part of gnome-bluetooth to BlueZ 5. The new API is a way
more simple than the old one.

One of the things that are gone with the BlueZ API is the Services
information, now we only rely on UUID to know which profiles are
supported. This doesn't affect us much since Services was already relying
UUID information.

The new API uses freedesktop DBus ObjectMananer and Properties, thus we
don't need to watch on many different interfaces for signals, for
example.

https://bugzilla.gnome.org/show_bug.cgi?id=685717