agent: Port agent code to BlueZ 5
The biggest difference of the new Agent API is that BlueZ 5 doesn't
register one agent per adapter anymore. There is now only one agent, to
handle all adapters at the same time and most of changes in this commit
are related to that.

The new API also introduces a new Agent method, RequestAuthorization(), to
give power to the user allow/deny JustWorks pairing(when both sides
doesn't require user interaction), so now bluetooth_agent_set_authorize_func
is related to that method and there is new
bluetooth_agent_set_authorize_service_func which is sets the callback to
the function to authorize services to trying to connect.

https://bugzilla.gnome.org/show_bug.cgi?id=685717