wizard: make the search page an intro page
It was previously a content page, which was fine except
when pairing failed and the Retry button was pressed.
That sent us back to the search page but a 'Go Back' button
was added, which made no sense and was crashing the wizard.
So make the page an intro page to avoid the Back button.