lib: Make default adapter work properly again
BlueZ has removed the DefaultAdapter concept so now we use the first
adapter plugged as the default adapter.

In case the default adapter is removed we iter to find a new adapter
to replace as default adapter, in case it exists we call
default_adapter_changed() to setup it as the new default.

We were able to get rid of some loops in the code in favor to a better
implementation using get_iter_from_path() to get the desired iter.

https://bugzilla.gnome.org/show_bug.cgi?id=701399