wizard: Work-around remote-initiated pairing race
When the TomTom Remote is trying to pair with us, the management
event with callback pin_code_request_callback() is getting called
before the mgmt event with callback connected_callback()
and the Class of device isn't set the first time we're asked for
a pincode.

Instead, wait for 1/2 second, as with a human user before replying
with the PIN code. This means that we'll get the connected event
processed, the class will be updated, and we can reply with the
actual PIN code we want to use.