build: Don't override Automake check: target
That causes warnings, which are obvious in the Continuous build view.
Instead, since the goal of this code is to execute xmllint, use a
stamp file target.

This fixes the warnings, and furthermore ensures that we only run
xmllint when the source file changes, instead of on every invocation
of make.

https://bugzilla.gnome.org/show_bug.cgi?id=721347