configure: require Gtk 3.11.2
bluetooth-settings-widget and bluetooth-settings-row both use
GtkListBoxRow, which was first introduced in GTK 3.9.4.
But since 4b2c4fd253bc49d2a, gtk_widget_set_margin_{start,end}
are also used, which means we need to depend on 3.11.2.

The variable $GTK_REQUIRED was actually not used in this file
so we need to change the PKG_CHECK_MODULES line to use it too.

Approved by Bastien Nocera <hadess@hadess.net> in IRC.
https://bugzilla.gnome.org/show_bug.cgi?id=722328