settings: Handle service authorisation requests for HID devices
When a HID device connects, we used to allow trusted or paired devices
to connect without any user interaction. But when setting up a PS3 pad
through cable pairing, the device would not be trusted yet.

In that case, show a dialogue to the user to allow them to choose
whether to allow that device to connect.

https://bugzilla.gnome.org/show_bug.cgi?id=752015