lib: Trust devices paired from the remote device
When pairing a device initiated from the remote device (a phone for
example), we don't receive any feedback that the pairing occurred. That
means that we won't get the option of setting the device as Trusted as
we would for a pairing initiated from this GNOME computer.

When a device is paired but not trusted and tries to connect to us,
assume that the device can connect to this service, otherwise we
wouldn't have paired it, right?

https://bugzilla.gnome.org/show_bug.cgi?id=763389