lib: Update UI for updated gnome-control-center UI
gnome-control-center panels are being given standardized layouts,
in order to provide a more seamless experience. This applies the
same layout rules to the Bluetooth Settings.

Apply a standard vertical spacing of 32px above and below
content and 12px below headings. Also aim for a standard row
height of 58px with the normal text size.

https://bugzilla.gnome.org/show_bug.cgi?id=786611