lib: Fix crash when exiting panel during pairing
When cancelling a pairing, make sure not to access user_data that might
not be valid anymore.

This code tried to access priv without it being set, as well as
trying to access before checking the result of the operation (which
might be "cancelled" meaning we shouldn't try to touch the UI).

 #0  0x00007f5ef0671ac7 in gtk_widget_destroy (widget=0xaaaaaaaaaaaaaaaa) at .../jhbuild/gtk+-3/gtk/gtkwidget.c:4719
 #1  0x00007f5ef0fd1ab3 in create_callback (source_object=0x2017120, res=0x205a250, user_data=0x1de3460) at .../jhbuild/gnome-bluetooth/lib/bluetooth-settings-widget.c:897
 #2  0x00007f5eeed8a9e8 in g_simple_async_result_complete (simple=0x205a250) at .../jhbuild/glib/gio/gsimpleasyncresult.c:801