build: Improve post-install script
The post install script is executing a rename process from when the
icons' filenames also contained their installation path, which is
not necessary anymore. It also contains a fixed path for datadir,
which might change.

This patch removes the unnecessary rename process and also uses
the provided datadir path, if the user provides one, or the default
path.

https://bugzilla.gnome.org/show_bug.cgi?id=790332