lib: Fix possible race when creating adapters or devices
When new D-Bus objects appear, listen to changes to their properties as
soon as possible to avoid missing out on events. This can happen if a
synchronous D-Bus call is made between the device appearing, and we
start listening to changes.

The window for that race is smaller now that we avoid creating
GDBusProxies and fetching properties synchronously when new adapters or
devices appear, but this close that window altogether.

https://bugzilla.gnome.org/show_bug.cgi?id=782530