lib: Remove spurious weak pointer reference to bluetooth agent
The bluetooth widget holds a single reference to the agent object. This
reference is unref'ed and cleared using g_object_cleared() in all relevant
cases.

This pointer should not be a weak reference. It will not cause issues
simply because agent will always be destroyed at the same time as the
widget.

https://bugzilla.gnome.org/show_bug.cgi?id=790720