client: Turn off discoverable when discovery finishes
Rather than turning discoverable on when the panel is launched, and off
when it exits (with all the problem that might entail during crashes),
use SetDiscoveryFilter's Discoverable property to make our adapter
discoverable during discovery. This will make sure that the discoverable
status is bound to the discovery process, limiting it in time, and
making sure that it stops when the settings go away, whatever the way it
does that.

Note that this also changes the meaning of the
"default-adapter-discoverable" property to mean "whether the default
adapter is visible by other devices during discovery", not whether the
default adapter is discoverable at that instant.