build: Improve GtkDoc documentation generation
Since gnome-bluetooth was ported to meson, version has acquired
a new parameter called `private_headers` which replaces the
`--ignore-headers` argument in `scan_args` parameter.
`--rebuild-types` argument is also not necessary because a types
does exist already.

`--sgml-mode` and `--output-format=xml` parameters are also not
necessary because meson already handles them automatically.

Finally, the `version.xml` is included now explicitly as a
content file because previously it was copied as a workaround.