Add BluetoothHdyColumn
This is a fork of HdyColumn introduced to fix a name collision in GNOME
Settings as is has both gnome-bluetooth's version of HdyColumn and its
own existing in parallel. It will be used in the next commit to replace
HdyColumn.

The widget couldn't be named BluetoothColumn as the name was already
taken.