client: Fix possible use-after-free
bluetooth_client_setup_device_finish() didn't copy the "device path"
string that came through attached to the result object, so any uses
after the async callback finished would have been from freed memory.

Always return a copy of the path instead, and make sure to free that
copy when we're done with it.