settings-widget: Really fix "don't show devices without a name"
The fix in commit 274b856f11bcbb2be67ebc25151c0ae796baab9a was overly
aggressive and did not create any rows for devices that appeared and
didn't have a name. The problem is that the rows were also not created
when that name appeared.

Always create rows for devices which we might end up showing but mark
them as invisible until a name is present.