settings-widget: Sort by time created rather than alphabetically
Sort devices by the time a row was created for them. This means that
newly discovered devices should appear at the bottom of the list, rather
than the top, making it less likely that the widget under the mouse or
finger would change before it's clicked.

This should also make the list of setup devices more static even if
the device names are changed.

Closes: #55