agent: Fix possible hang when switching Settings panels
There are usually 2 reasons why the agent might go away:
- we're switching panels in the Settings
- the Settings executable is exit()ing

In the former case, we can't actually do anything if unregistering the
agent fails, and we can already remove our object path without waiting
on anyone, so just print some debug if it fails.

In the latter case, we'll be gone from the bus before before the method
has time to answer.

This fixes possible hangs while unregistering the agent when switching
panels with some broken Bluetooth adapters and/or drivers.

Closes: #49