settings: Replace canberra-gtk by to gsound
Mainly to reduce the number of GTK3 dependencies before actually start
porting the library. Fortunately, the usage of libcanberra-gtk by
gnome-bluetooth is rather trivial, so porting to gsound is just a few
lines of code away.

Replace canberra-gtk by gsound, adjust the build system to reflect
that, and update the CI file to install gsound-devel instead of
libcanberra-devel.