meson: bump gettext package suffix
Currently, both gnome-bluetooth-1.0 and 2.0 are using "2"
as a suffix for the gettext locale files which causes them to
overalp. Looks like 2 has been used since the initial commit
of the meson port, back in 3.20 cycle, although not sure if
there is any special meaning to the number. Bump it anyway
so that translations won't overlap between the two versions.