lib: Add battery reporting through BluetoothDevice
Connect and monitor UPower devices for devices reporting their battery
status.

This makes sure that we can show battery status for devices
that emit their battery status through bluez (via the BATT LE service,
or the org.bluez.BatteryProviderManager1 API, as well as devices for
which the kernel is the battery info provider.

Closes: #102