lib: Merge device removal with adapter removal
When an adapter is removed while bluetoothd is still running (eg. not
crashing as we also want to handle), suppress the emission of
device-removed so that front-ends don't think that every device is
removed.

We implement this by queue device-removed signals until the next
mainloop iteration instead of sending it straight away.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/1426
Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/5058