lib: Export a more precise adapter power state
Unfortunately, powering on/off Bluetooth adapters has become longer,
and less reliable over time, so front-ends need more information to
be able to figure out what it happening.

As a first pass, export whether the adapter is in the process of being
turned on, or turned off, based on our own request.