lib: Implement default-adapter-state property using bluez
Use new adapter property in bluez to implement the
default-adapter-state. It should work better, as it doesn't rely only on
internal BluetoothClient state in a single process, and can get
information when bluetoothd is powering up the adapter.

Note that this property still works in the absence of a new enough
bluetoothd, without the transitional states.

Closes: #121