pairing-dialog: Fix pairing with keyboard with typing notification
The invisible character we fetch from an entry might be zero if it
isn't set, for example when the entry isn't setup to hide text. This
would result in the passkey that's displayed being truncated as the
invisible character would be a nul terminator.

Make sure we have a valid invisible character, so the passkey doesn't
disappear after the first typed character when pairing a keyboard with
typing notification.

Closes: #125