settings: Sync state and active
The internal state of the switch for a device's connection sometimes can
be wrong. This happens when the state and active properties of the
switch are not in sync. This is reflected visually when the toggle is
not active, but blue, or toggle is active, but gray.

This happens because of a change in gtk
https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/5442.

To fix it we have to make sure that

- In case of a success we set the `state` of the switch to the target
  value given by the state-set signal
- In case of a failure, the switch has to go back to its previous
  position, so we set `active` to the previous value accordingly

At the start of state-set we return early if the target state already
matches the current active, so we set both properties at once to be
extra sure they are synced.

Fixes: https://gitlab.gnome.org/GNOME/gnome-bluetooth/-/issues/132