client: Remove device before emitting signal
Otherwise any code that uses the items from the devices model
in response to the signal will still get the removed device.

While there are ways to deal with this on the caller side, it is
less confusing when the model has already been updated at the time
the signal is emitted. For the `device-added` signal this is already
the case, as the item is added before emitting the signal.