client: Fix name owner string leaks
g_dbus_proxy_get_name_owner() returned value is "transfer full"
unlike g_dbus_proxy_get_interface_name() and
g_dbus_proxy_get_object_path() which are "transfer none".

Spotted while valgrinding gnome-control-center.

Instead of just fixing the leak we completely remove
the use of the affected "name" variable, because the
idea for it was to be able to track bluetoothd
appearing and disappearing, but we already have the
bluetooth-agent.c code doing that.

5 bytes in 1 blocks are definitely lost in loss record 223 of 29,728
   at 0x483EB26: malloc (vg_replace_malloc.c:446)
   by 0x4B8C478: g_malloc (gmem.c:100)
   by 0x4B8C755: g_malloc_n (gmem.c:410)
   by 0x4BB1F51: g_strdup (gstrfuncs.c:323)
   by 0x49A4F88: g_strdup_inline (gstrfuncs.h:321)
   by 0x49A4F88: g_dbus_proxy_get_name_owner (gdbusproxy.c:2309)
   by 0x6A161EB: adapter_added.isra.0 (bluetooth-client.c:692)
   by 0x6A1650A: object_manager_new_callback (bluetooth-client.c:925)
   by 0x4907920: g_task_return_now (gtask.c:1363)
   by 0x490798C: complete_in_idle_cb (gtask.c:1377)
   by 0x4B83D8D: g_idle_dispatch (gmain.c:6411)
   by 0x4B804A9: g_main_dispatch (gmain.c:3524)
   by 0x4B818D9: g_main_context_dispatch_unlocked (gmain.c:4375)
   by 0x4B81A99: g_main_context_iterate_unlocked (gmain.c:4440)
   by 0x4B81BC4: g_main_context_iteration (gmain.c:4505)
   by 0x494CF2E: g_application_run (gapplication.c:2715)
   by 0x41511C: main (main.c:47)